# openapi.yaml
openapi: 3.0.3
info:
  title: DuoStackDemo API
  description: |
    API specification for DuoStackDemo.
    This spec serves as the single source of truth for both Node.js and .NET backends.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/v1
    description: Node.js backend
  - url: http://localhost:5000/api/v1
    description: .NET backend
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: refreshToken
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: string

    # Enums
...
    Room:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        isPublic:
          type: boolean
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
    CreateRoomRequest:
      type: object
      required:
        - name
        - isPublic
      properties:
        name:
          type: string
          minLength: 1
        isPublic:
          type: boolean
    UpdateRoomRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        isPublic:
          type: boolean

    # Page (CMS) schemas
...
paths:
...
  /rooms:
    get:
      tags:
        - Rooms
      summary: Get all rooms
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Rooms retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Room'

    post:
      tags:
        - Rooms
      summary: Create room (Admin+ only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoomRequest'
      responses:
        '201':
          description: Room created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  data:
                    $ref: '#/components/schemas/Room'
        '403':
          description: Forbidden - Admin role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /rooms/{id}:
    put:
      tags:
        - Rooms
      summary: Update room (Admin+ only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoomRequest'
      responses:
        '200':
          description: Room updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  data:
                    $ref: '#/components/schemas/Room'
        '403':
          description: Forbidden - Admin role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Rooms
      summary: Soft delete room (Admin+ only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Room deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  data:
                    type: object
                    properties:
                      message:
                        type: string
        '403':
          description: Forbidden - Admin role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Page (CMS) endpoints
...